requirements.txt
opencv-python==4.1.1.26
opencv-contrib-python==4.2.0.34
opencv-python-headless==4.1.1.26
numpy==1.18.0
flask==1.1.1
pillow==6.2.0


